add loop over int to template funcs
Go templates can loop over slices, maps and channels, but as far as I know are
missing the ability to simply specify loop iterations as an integer. This
function adds that ability by generating a slice of `struct{}` to loop over.
Because `struct{}` takes up no space, this causes no memory allocation.

This functionality is useful (at the very least) to specify a variable number of
instances to create.