add cf value of "current" stack to template vars
This gives templates a way to reference the parameters under "cf" in a stack the
config file without referencing a stack name directly. The "currentstack" key in
the map passed to the template will contain the "cf" map of the stack which is
being generated.

In this way, a single template can be used by multiple different stacks and have
different outputs.